# api_docs.json
{
"openapi": "3.0.0",
"info": {
    "title": "vending_machine",
    "version": "1.0.0"
},
"servers": [
    {
        "url": "http://{{base_url}}"
    }
],
"paths": {
...
    "/product/": {
        "post": {
            "tags": [
            "default"
            ],
            "summary": "create product",
            "requestBody": {
            "content": {
                "application/json": {
                    "schema": {
                        "type": "object",
                        "example": {
                        "productName": "pepsi",
                        "cost": 30,
                        "amountAvailable": 8
                        }
                    }
                }
            }
            },
            "responses": {
            "201": {
                "description": "Created",
                "headers": {
                    "Content-Security-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                        }
                    },
                    "Cross-Origin-Embedder-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "require-corp"
                        }
                    },
                    "Cross-Origin-Opener-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "Cross-Origin-Resource-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "X-DNS-Prefetch-Control": {
                        "schema": {
                        "type": "string",
                        "example": "off"
                        }
                    },
                    "X-Frame-Options": {
                        "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                        }
                    },
                    "Strict-Transport-Security": {
                        "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                        }
                    },
                    "X-Download-Options": {
                        "schema": {
                        "type": "string",
                        "example": "noopen"
                        }
                    },
                    "X-Content-Type-Options": {
                        "schema": {
                        "type": "string",
                        "example": "nosniff"
                        }
                    },
                    "Origin-Agent-Cluster": {
                        "schema": {
                        "type": "string",
                        "example": "?1"
                        }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                        "schema": {
                        "type": "string",
                        "example": "none"
                        }
                    },
                    "Referrer-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "no-referrer"
                        }
                    },
                    "X-XSS-Protection": {
                        "schema": {
                        "type": "integer",
                        "example": "0"
                        }
                    },
                    "Access-Control-Allow-Origin": {
                        "schema": {
                        "type": "string",
                        "example": "*"
                        }
                    },
                    "Content-Type": {
                        "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                        }
                    },
                    "Content-Length": {
                        "schema": {
                        "type": "integer",
                        "example": "486"
                        }
                    },
                    "ETag": {
                        "schema": {
                        "type": "string",
                        "example": "W/\"1e6-jisrD3dtzCvPMvSNyWCLeTebwJc\""
                        }
                    },
                    "Set-Cookie": {
                        "schema": {
                        "type": "string",
                        "example": "vending_machine=s%3AsD674G98U3L9Ln9JUhtHBw3BOMSIAk_7.34rlMvtYP7YQAp6Nvozf%2FBs3e6VeBfCwreVXJtARDj8; Path=/; Expires=Fri, 31 Mar 2023 12:47:07 GMT; HttpOnly"
                        }
                    },
                    "Date": {
                        "schema": {
                        "type": "string",
                        "example": "Fri, 31 Mar 2023 12:42:07 GMT"
                        }
                    },
                    "Connection": {
                        "schema": {
                        "type": "string",
                        "example": "keep-alive"
                        }
                    },
                    "Keep-Alive": {
                        "schema": {
                        "type": "string",
                        "example": "timeout=5"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                        "type": "object"
                        },
                        "example": {
                        "status": "success",
                        "message": "Product created succesfully",
                        "data": {
                            "productName": "coke",
                            "amountAvailable": 12,
                            "cost": 15,
                            "sellerID": "64266dd2a76abb80cdc2ea69",
                            "createdAt": "2023-03-31T12:42:07.618Z",
                            "updatedAt": "2023-03-31T12:42:07.618Z",
                            "seller": {
                                "username": "hephhay",
                                "role": "seller",
                                "deposit": {
                                    "5": 0,
                                    "10": 0,
                                    "20": 0,
                                    "50": 0,
                                    "100": 0
                                },
                                "createdAt": "2023-03-31T05:21:22.877Z",
                                "updatedAt": "2023-03-31T08:59:33.036Z",
                                "balance": 0,
                                "id": "64266dd2a76abb80cdc2ea69"
                            },
                            "id": "6426d51fb7144971d70b7ecc"
                        }
                        }
                    }
                }
            },
            "400": {
                "description": "Bad Request",
                "headers": {
                    "Content-Security-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                        }
                    },
                    "Cross-Origin-Embedder-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "require-corp"
                        }
                    },
                    "Cross-Origin-Opener-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "Cross-Origin-Resource-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "X-DNS-Prefetch-Control": {
                        "schema": {
                        "type": "string",
                        "example": "off"
                        }
                    },
                    "X-Frame-Options": {
                        "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                        }
                    },
                    "Strict-Transport-Security": {
                        "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                        }
                    },
                    "X-Download-Options": {
                        "schema": {
                        "type": "string",
                        "example": "noopen"
                        }
                    },
                    "X-Content-Type-Options": {
                        "schema": {
                        "type": "string",
                        "example": "nosniff"
                        }
                    },
                    "Origin-Agent-Cluster": {
                        "schema": {
                        "type": "string",
                        "example": "?1"
                        }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                        "schema": {
                        "type": "string",
                        "example": "none"
                        }
                    },
                    "Referrer-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "no-referrer"
                        }
                    },
                    "X-XSS-Protection": {
                        "schema": {
                        "type": "integer",
                        "example": "0"
                        }
                    },
                    "Access-Control-Allow-Origin": {
                        "schema": {
                        "type": "string",
                        "example": "*"
                        }
                    },
                    "Content-Type": {
                        "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                        }
                    },
                    "Content-Length": {
                        "schema": {
                        "type": "integer",
                        "example": "319"
                        }
                    },
                    "ETag": {
                        "schema": {
                        "type": "string",
                        "example": "W/\"13f-KZ4QfMtPlHBKw2VUJHK9jJnBwnE\""
                        }
                    },
                    "Set-Cookie": {
                        "schema": {
                        "type": "string",
                        "example": "vending_machine=s%3Ajt3MbH1IhYACBBcO1qGHAjvm9vxKLO3o.99Kd6GxqNEIpxx7zvXRmtUglfu1CaguphZDXL0H5dMs; Path=/; Expires=Fri, 31 Mar 2023 12:31:13 GMT; HttpOnly"
                        }
                    },
                    "Date": {
                        "schema": {
                        "type": "string",
                        "example": "Fri, 31 Mar 2023 12:26:13 GMT"
                        }
                    },
                    "Connection": {
                        "schema": {
                        "type": "string",
                        "example": "keep-alive"
                        }
                    },
                    "Keep-Alive": {
                        "schema": {
                        "type": "string",
                        "example": "timeout=5"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                        "type": "object"
                        },
                        "example": {
                        "status": "failed",
                        "message": "Invalid Input",
                        "details": [
                            {
                                "code": "too_small",
                                "minimum": 0,
                                "type": "number",
                                "inclusive": true,
                                "exact": false,
                                "message": "Number must be greater than or equal to 0",
                                "path": [
                                    "amountAvailable"
                                ]
                            },
                            {
                                "code": "not_multiple_of",
                                "multipleOf": 5,
                                "message": "Number must be a multiple of 5",
                                "path": [
                                    "cost"
                                ]
                            }
                        ]
                        }
                    }
                }
            }
            }
        }
    },
...
    "/product": {
        "get": {
            "tags": [
            "default"
            ],
            "summary": "get products",
            "parameters": [
            {
                "name": "amountAvailable_gte",
                "in": "query",
                "schema": {
                    "type": "integer"
                },
                "example": "0"
            },
            {
                "name": "amountAvailable_lte",
                "in": "query",
                "schema": {
                    "type": "integer"
                },
                "example": "100"
            },
            {
                "name": "cost_gte",
                "in": "query",
                "schema": {
                    "type": "integer"
                },
                "example": "0"
            },
            {
                "name": "cost_lte",
                "in": "query",
                "schema": {
                    "type": "integer"
                },
                "example": "50"
            },
            {
                "name": "productName",
                "in": "query",
                "schema": {
                    "type": "string"
                },
                "example": "e"
            },
            {
                "name": "amountAvailable",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            {
                "name": "cost",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            {
                "name": "productID",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            }
            ],
            "responses": {
            "200": {
                "description": "OK",
                "headers": {
                    "Content-Security-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                        }
                    },
                    "Cross-Origin-Embedder-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "require-corp"
                        }
                    },
                    "Cross-Origin-Opener-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "Cross-Origin-Resource-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "same-origin"
                        }
                    },
                    "X-DNS-Prefetch-Control": {
                        "schema": {
                        "type": "string",
                        "example": "off"
                        }
                    },
                    "X-Frame-Options": {
                        "schema": {
                        "type": "string",
                        "example": "SAMEORIGIN"
                        }
                    },
                    "Strict-Transport-Security": {
                        "schema": {
                        "type": "string",
                        "example": "max-age=15552000; includeSubDomains"
                        }
                    },
                    "X-Download-Options": {
                        "schema": {
                        "type": "string",
                        "example": "noopen"
                        }
                    },
                    "X-Content-Type-Options": {
                        "schema": {
                        "type": "string",
                        "example": "nosniff"
                        }
                    },
                    "Origin-Agent-Cluster": {
                        "schema": {
                        "type": "string",
                        "example": "?1"
                        }
                    },
                    "X-Permitted-Cross-Domain-Policies": {
                        "schema": {
                        "type": "string",
                        "example": "none"
                        }
                    },
                    "Referrer-Policy": {
                        "schema": {
                        "type": "string",
                        "example": "no-referrer"
                        }
                    },
                    "X-XSS-Protection": {
                        "schema": {
                        "type": "integer",
                        "example": "0"
                        }
                    },
                    "Access-Control-Allow-Origin": {
                        "schema": {
                        "type": "string",
                        "example": "*"
                        }
                    },
                    "Content-Type": {
                        "schema": {
                        "type": "string",
                        "example": "application/json; charset=utf-8"
                        }
                    },
                    "Content-Length": {
                        "schema": {
                        "type": "integer",
                        "example": "894"
                        }
                    },
                    "ETag": {
                        "schema": {
                        "type": "string",
                        "example": "W/\"37e-tHkNmkeP6k5javqbppcjdUTO+VI\""
                        }
                    },
                    "Set-Cookie": {
                        "schema": {
                        "type": "string",
                        "example": "vending_machine=s%3ATp8A43hrPVlzMBSky7-dEqdf2amTl2ue.LSmJdeXWp0xxirk%2BuNq5MmXg%2FsER7xau3cCiCGvHAu8; Path=/; Expires=Fri, 31 Mar 2023 19:11:28 GMT; HttpOnly"
                        }
                    },
                    "Date": {
                        "schema": {
                        "type": "string",
                        "example": "Fri, 31 Mar 2023 19:06:28 GMT"
                        }
                    },
                    "Connection": {
                        "schema": {
                        "type": "string",
                        "example": "keep-alive"
                        }
                    },
                    "Keep-Alive": {
                        "schema": {
                        "type": "string",
                        "example": "timeout=5"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                        "type": "object"
                        },
                        "example": {
                        "status": "success",
                        "message": "products found",
                        "data": [
                            {
                                "productName": "coke",
                                "amountAvailable": 12,
                                "cost": 15,
                                "sellerID": "64266dd2a76abb80cdc2ea69",
                                "createdAt": "2023-03-31T13:13:47.647Z",
                                "updatedAt": "2023-03-31T13:13:47.647Z",
                                "seller": {
                                    "username": "hephhay",
                                    "role": "seller",
                                    "deposit": {
                                    "5": 0,
                                    "10": 0,
                                    "20": 0,
                                    "50": 0,
                                    "100": 0
                                    },
                                    "createdAt": "2023-03-31T05:21:22.877Z",
                                    "updatedAt": "2023-03-31T08:59:33.036Z",
                                    "balance": 0,
                                    "id": "64266dd2a76abb80cdc2ea69"
                                },
                                "id": "6426dc8b137e0b9dac606ac8"
                            },
                            {
                                "productName": "pepsi",
                                "amountAvailable": 8,
                                "cost": 30,
                                "sellerID": "64266dd2a76abb80cdc2ea69",
                                "createdAt": "2023-03-31T13:18:13.037Z",
                                "updatedAt": "2023-03-31T13:18:13.037Z",
                                "seller": {
                                    "username": "hephhay",
                                    "role": "seller",
                                    "deposit": {
                                    "5": 0,
                                    "10": 0,
                                    "20": 0,
                                    "50": 0,
                                    "100": 0
                                    },
                                    "createdAt": "2023-03-31T05:21:22.877Z",
                                    "updatedAt": "2023-03-31T08:59:33.036Z",
                                    "balance": 0,
                                    "id": "64266dd2a76abb80cdc2ea69"
                                },
                                "id": "6426dd95140eef0a01ea7a84"
                            }
                        ]
                        }
                    }
                }
            }
            }
        }
    },
...
}
}